requirements.txt
blinker==1.6.2
boto3==1.26.137
botocore==1.29.137
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.3.2
Flask-Login==0.6.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.30.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.15
Werkzeug==2.3.4
zipp==3.15.0
